+++ src/test/java/com/firestationquiz/gamedataservice/GameDataServiceTests.java
@@ -1,0 +1,18 @@
+package com.firestationquiz.gamedataservice;
+
+import com.firestationquiz.gamedataservice.models.Station;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+
+@ActiveProfiles("test")
+@SpringBootTest()
+public class GameDataServiceTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
